% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_common_ids.R
\name{add_common_ids}
\alias{add_common_ids}
\title{RM2C2: Scoring, Summarizing}
\usage{
add_common_ids(df, cols = NA, verbose = F)
}
\arguments{
\item{df}{class: data frame}

\item{cols}{class: named list}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
add_common_ids(df, cols=list(burst="1"))
}
\keyword{cognition}
\keyword{m2c2,}
