% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_md5_hash_col.R
\name{add_md5_hash_col}
\alias{add_md5_hash_col}
\title{RM2C2: Scoring, Summarizing}
\usage{
add_md5_hash_col(data, algo = "md5")
}
\arguments{
\item{data}{class: data.frame}

\item{algo}{class: string}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
add_md5_hash_col(data, algo="md5")
}
