% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadate_cols.R
\name{add_metadate_cols}
\alias{add_metadate_cols}
\title{RM2C2: Scoring, Summarizing}
\usage{
add_metadate_cols(df, time_var = "timestamp", time_format = "\%m/\%d/\%Y")
}
\arguments{
\item{df}{class: dataframe}

\item{time_var}{class: string}

\item{time_format}{class: string}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
add_metadate_cols(df)
}
\keyword{cognition}
\keyword{m2c2,}
