% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to.R
\name{add_to}
\alias{add_to}
\title{RM2C2: Scoring, Summarizing}
\usage{
add_to(x, add = 1)
}
\arguments{
\item{x}{class: numeric; original number}

\item{add}{class: numeric; number to add}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
add_to(x1,x2,y1,y2)
}
\keyword{cognition}
\keyword{m2c2,}
