% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_colname_prefix.R
\name{append_colname_prefix}
\alias{append_colname_prefix}
\title{RM2C2dev}
\usage{
append_colname_prefix(data, group_var, var_prefix)
}
\description{
RM2C2dev
}
\examples{
append_colname_prefix(data, group_var, var_prefix)
}
