% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_metadate_cols.R
\name{append_metadate_cols}
\alias{append_metadate_cols}
\title{RM2C2_dev: Scoring, Summarizing}
\usage{
append_metadate_cols(
  data,
  time_var = "timestamp",
  time_format = "\%m/\%d/\%Y",
  time_origin = "1970-01-01",
  timezone = Sys.timezone(),
  append_rounded = T
)
}
\arguments{
\item{time_var}{class: string}

\item{time_format}{class: string}

\item{df}{class: dataframe}
}
\description{
RM2C2_dev: Scoring, Summarizing
}
\examples{
append_metadate_cols(df)
}
