% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep_schedule_from_csv.R
\name{beep_schedule_from_csv}
\alias{beep_schedule_from_csv}
\title{RM2C2dev}
\usage{
beep_schedule_from_csv(
  filename = NA,
  schedule = NA,
  schedule_public_label = "default"
)
}
\description{
RM2C2dev
}
