% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_launch_tag.R
\name{build_launch_tag}
\alias{build_launch_tag}
\title{RM2C2: Scoring, Summarizing}
\usage{
build_launch_tag(
  dolphin,
  launch,
  sensors,
  settings,
  root_tag,
  launch_tag,
  sensor_tag,
  settings_tag
)
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
build_launch_tag(df, root_tag)
}
\keyword{cognition}
\keyword{m2c2,}
