% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_menu_tag.R
\name{build_menu_tag}
\alias{build_menu_tag}
\title{RM2C2: Scoring, Summarizing}
\usage{
build_menu_tag(root, menu, menu_tag, title_tag, element_tag, action_tag)
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
build_menu_tag(df, root_tag)
}
\keyword{cognition}
\keyword{m2c2,}
