% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pack_tag.R
\name{build_pack_tag}
\alias{build_pack_tag}
\title{RM2C2: Scoring, Summarizing}
\usage{
build_pack_tag(
  root,
  pack,
  pack_tag,
  game_tag,
  screen_tag,
  branch_tag,
  nextscreen_tag,
  element_tag,
  option_tag
)
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
build_pack_tag(df, root_tag)
}
\keyword{cognition}
\keyword{m2c2,}
