% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_root_tag.R
\name{build_root_tag}
\alias{build_root_tag}
\title{RM2C2: Scoring, Summarizing}
\usage{
build_root_tag(df, root_tag)
}
\arguments{
\item{df}{class: string}

\item{root_tag}{class: string}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
build_root_tag(df, root_tag)
}
\keyword{cognition}
\keyword{m2c2,}
