% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_report.R
\name{compliance_report}
\alias{compliance_report}
\title{RM2C2: Scoring, Summarizing}
\usage{
compliance_report(df, ...)
}
\arguments{
\item{df}{class: numeric; original X}

\item{...}{class string, vector}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
compliance_report(df)
}
\keyword{cognition}
\keyword{m2c2,}
