% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{RM2C2: Scoring, Summarizing}
\usage{
distance(x1, x2, y1, y2)
}
\arguments{
\item{x1}{class: numeric; original X}

\item{x2}{class: numeric; moved to X}

\item{y1}{class: numeric; original Y}

\item{y2}{class: numeric; moved to Y}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
distance(x1,x2,y1,y2)
}
\keyword{cognition}
\keyword{m2c2,}
