% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_server_zip.R
\name{download_server_zip}
\alias{download_server_zip}
\title{RM2C2dev}
\usage{
download_server_zip(
  url,
  params,
  save_filename,
  overwrite_zip = F,
  unzip = T,
  remove_zip = F
)
}
\arguments{
\item{url}{class: string}

\item{params}{class: list}

\item{save_filename}{class: string}

\item{overwrite_zip}{class: boolean}

\item{unzip}{class: boolean}

\item{remove_zip}{class: boolean}
}
\description{
RM2C2dev
}
\examples{
download_server_zip(url, params, save_filename, overwrite_zip=T, unzip=T, remove_zip=F)
}
