% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_synapse_cogtask_table_all.R,
%   R/download_synapse_table_all.R
\name{download_synapse_table_all}
\alias{download_synapse_table_all}
\alias{download_synapse_cogtask_table_all}
\title{RM2C2dev}
\usage{
download_synapse_cogtask_table_all(synapse_email = NA, synapse_pw = NA,
  synapse_id = NA, uuids = NA)

download_synapse_table_all(synapse_email = NA, synapse_pw = NA,
  synapse_id = NA)
}
\arguments{
\item{synapse_email}{class: string}

\item{synapse_pw}{class: string}

\item{synapse_id}{class: string}

\item{synapse_email}{class: string}

\item{synapse_pw}{class: string}

\item{synapse_id}{class: string}
}
\description{
RM2C2dev

RM2C2dev
}
\examples{
download_synapse_table_all(synapse_email = NA, synapse_pw = NA, synapse_id = NA, record_uuids = NA)
login to synapse
build the query
get raw, unmanipulated dataframe matching query results
download_synapse_table_all(synapse_email = NA, synapse_pw = NA, synapse_id = NA, record_uuids = NA)
login to synapse
build the query
get raw, unmanipulated dataframe matching query results
}
