% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_synapse_table_from_start_date_plus.R
\name{download_synapse_table_from_start_date_plus}
\alias{download_synapse_table_from_start_date_plus}
\title{RM2C2dev}
\usage{
download_synapse_table_from_start_date_plus(
  synapse_email = NA,
  synapse_pw = NA,
  synapse_id = NA,
  start_date = NA,
  search_op = ">="
)
}
\arguments{
\item{synapse_email}{class: string}

\item{synapse_pw}{class: string}

\item{synapse_id}{class: string}

\item{start_date}{class: string}

\item{search_op}{class: string}
}
\description{
RM2C2dev
}
\examples{
download_synapse_table_from_start_date_plus(synapse_email = synapse_email, synapse_pw =  synapse_pw, synapse_id = synapse_survey_data_table_id, start_date = "'2020-05-01'", search_op = ">=")
login to synapse
build the query
get raw, unmanipulated dataframe matching query results
}
