% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_zip_server.R
\name{download_zip_server}
\alias{download_zip_server}
\title{RM2C2: Scoring, Summarizing}
\usage{
download_zip_server(url, params, save_filename, overwrite_zip = F,
  unzip = T, remove_zip = F)
}
\arguments{
\item{url}{class: string}

\item{params}{class: list}

\item{save_filename}{class: string}

\item{overwrite_zip}{class: boolean}

\item{unzip}{class: boolean}

\item{remove_zip}{class: boolean}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
download_zip_server(url, params, save_filename, overwrite_zip=T, unzip=T, remove_zip=F)
}
