% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cog_duplicate_records.R
\name{filter_cog_duplicate_records}
\alias{filter_cog_duplicate_records}
\title{RM2C2: Scoring, Summarizing}
\usage{
filter_cog_duplicate_records(df)
}
\arguments{
\item{df}{class: numeric; original X}

\item{group_vars}{class: vector; factors to group data by}

\item{time_var}{class: vector; variable for time}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
filter_cog_duplicate_records(df)
}
\keyword{cognition}
\keyword{m2c2,}
