% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date.R
\name{filter_date}
\alias{filter_date}
\title{RM2C2: Scoring, Summarizing}
\usage{
filter_date(df, date_col = "date", start_day = "", end_day = "")
}
\arguments{
\item{df}{class: numeric; original X}

\item{date_col}{class: numeric; moved to X}

\item{start_day}{class: numeric; moved to X}

\item{end_day}{class: numeric; moved to X}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
filter_date(df,date_col='date', start_day="2019-02-01", end_day="2019-03-01")
}
\keyword{cognition}
\keyword{m2c2,}
