% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_exact_duplicates.R
\name{filter_exact_duplicates}
\alias{filter_exact_duplicates}
\title{RM2C2: Scoring, Summarizing}
\usage{
filter_exact_duplicates(df, group_vars = c("user_id", "Session"))
}
\arguments{
\item{df}{class: numeric; original X}

\item{group_vars}{class: vector; factors to group data by}

\item{time_var}{class: vector; variable for time}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
filter_exact_duplicates(df, group_vars=c("user_id", "Session"))
}
\keyword{cognition}
\keyword{m2c2,}
