% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_RT.R
\name{filter_RT}
\alias{filter_RT}
\title{RM2C2: Scoring, Summarizing}
\usage{
filter_RT(df, RT_col = "response_time", lower_RT = 200, upper_RT = 10000)
}
\arguments{
\item{df}{class: numeric; original X}

\item{RT_col}{class: string}

\item{lower_RT}{class: numeric; moved to X}

\item{upper_RT}{class: numeric; moved to X}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
filter_RT(df, RT_col = 'response_time', lower_RT = 200, upper_RT = 10000)
}
\keyword{cognition}
\keyword{m2c2,}
