% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_studyday.R
\name{filter_studyday}
\alias{filter_studyday}
\title{RM2C2: Scoring, Summarizing}
\usage{
filter_studyday(df, study_col = "studyday", start_day = 1, end_day = 14)
}
\arguments{
\item{df}{class: numeric; original X}

\item{study_col}{class: string}

\item{start_day}{class: numeric; moved to X}

\item{end_day}{class: numeric; moved to X}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
filter_studyday(df,study_col='date', start_day="2019-02-01", end_day="2019-03-01")
}
\keyword{cognition}
\keyword{m2c2,}
