% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates_keeplast.R,
%   R/filter_survey_duplicate_records.R
\name{filter_survey_duplicate_records}
\alias{filter_survey_duplicate_records}
\title{RM2C2: Scoring, Summarizing}
\usage{
filter_survey_duplicate_records(
  df,
  group_vars = c("user_id", "Session"),
  time_var = "End Time"
)

filter_survey_duplicate_records(
  df,
  group_vars = c("user_id", "Session"),
  time_var = "End Time"
)
}
\arguments{
\item{df}{class: numeric; original X}

\item{group_vars}{class: vector; factors to group data by}

\item{time_var}{class: vector; variable for time}
}
\description{
RM2C2: Scoring, Summarizing

RM2C2: Scoring, Summarizing
}
\examples{
filter_duplicates_keeplast(df, group_vars=c("user_id", "Session"), time_var="End Time")
filter_survey_duplicate_records(df, group_vars=c("user_id", "Session"), time_var="End Time")
}
\keyword{cognition}
\keyword{m2c2,}
