% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codebooks_from_dir.R
\name{generate_codebooks_from_dir}
\alias{generate_codebooks_from_dir}
\title{RM2C2dev}
\usage{
generate_codebooks_from_dir(dir, codebook_type = "simple", write_file = T)
}
\arguments{
\item{dir}{class: string}

\item{codebook_type}{class: string}
}
\description{
RM2C2dev
}
\examples{
generate_codebooks_from_dir('data/')
}
