% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_daily_report.R
\name{generate_daily_report}
\alias{generate_daily_report}
\title{RM2C2dev}
\usage{
generate_daily_report(
  pack_list,
  ignore_packs = NULL,
  by_pack = F,
  plot = F,
  debug = F,
  app_version = 1.3
)
}
\description{
RM2C2dev
}
