% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_study_xml.R
\name{generate_study_xml}
\alias{generate_study_xml}
\title{RM2C2: Scoring, Summarizing}
\usage{
generate_study_xml(input_file, project_name, verbose = T)
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
generate_study_xml(input_file, project_name, verbose=T)
}
\keyword{cognition}
\keyword{m2c2,}
