% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m2c2_auth_token.R
\name{get_m2c2_auth_token}
\alias{get_m2c2_auth_token}
\title{RM2C2dev}
\usage{
get_m2c2_auth_token(
  server_url = "",
  user_id = "",
  password = "",
  study_id = ""
)
}
\arguments{
\item{server_url}{class: string}

\item{user_id}{class: string}

\item{password}{class: string}

\item{study_id}{class: string}
}
\description{
RM2C2dev
}
\examples{
get_m2c2_auth_token(server_url = 'https://m2c2.survey.psu.edu/test/', user_id = 'r_user', password = 'test', study_id = 'test')
}
