% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_src_data_brick.R
\name{get_src_data_brick}
\alias{get_src_data_brick}
\title{RM2C2dev}
\usage{
get_src_data_brick(
  study_id = NA,
  server_url = NA,
  study_code = NA,
  user_id = NA,
  user_pass = NA,
  packs = NA,
  app_version = 1.3
)
}
\description{
RM2C2dev
}
