% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synapse_tables.R
\name{get_synapse_tables}
\alias{get_synapse_tables}
\title{RM2C2dev}
\usage{
get_synapse_tables(synapse_email = NA, synapse_pw = NA,
  synapse_project_id = NA)
}
\arguments{
\item{synapse_email}{class: string}

\item{synapse_pw}{class: string}

\item{synapse_project_id}{class: string}
}
\description{
RM2C2dev
}
\examples{
get_synapse_tables(synapse_email = NA, synapse_pw = NA, synapse_project_id = NA)
login to synapse
load all objects
type convert to list
create blank data.frame for merging
}
