% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_datetime.R
\name{get_tidy_datetime}
\alias{get_tidy_datetime}
\title{RM2C2dev}
\usage{
get_tidy_datetime(delim = "_")
}
\arguments{
\item{datetime}{class: POSIXct}

\item{timezone}{class: string}
}
\description{
RM2C2dev
}
\examples{
get_tidy_datetime(datetime=NA, timezone="UTC")
}
