% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_values.R
\name{get_unique_values}
\alias{get_unique_values}
\title{RM2C2dev}
\usage{
get_unique_values(data)
}
\arguments{
\item{col}{class: vector}
}
\description{
RM2C2dev
}
\examples{
get_unique_values(data$dates)
}
