% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_frame_tibble.R
\name{is_data_frame_tibble}
\alias{is_data_frame_tibble}
\title{RM2C2dev}
\usage{
is_data_frame_tibble(data)
}
\arguments{
\item{data}{class: data.frame}
}
\description{
RM2C2dev
}
\examples{
is_data_frame_tibble(data)
check if data is tibble or data.frame
}
