% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_beep_schedule.R
\name{jitter_beep_schedule}
\alias{jitter_beep_schedule}
\title{RM2C2dev}
\usage{
jitter_beep_schedule(
  schedule = NA,
  schedule_name = "default",
  jitter_time_by_hours = 1,
  jitter_days_by_ndays = 0
)
}
\description{
RM2C2dev
}
