% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_m2c2_study_xml.R
\name{make_m2c2_study_xml}
\alias{make_m2c2_study_xml}
\title{RM2C2dev}
\usage{
make_m2c2_study_xml(
  csv_file = "",
  menu_str = "",
  study_name = "",
  study_servername = "",
  study_wave = "",
  study_server = "https://m2c2.survey.psu.edu/eas",
  app_version = "1.3",
  debug = 0
)
}
\description{
RM2C2dev
}
