% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_datetime_filename.R
\name{make_tidy_datetime_filename}
\alias{make_tidy_datetime_filename}
\title{RM2C2: Scoring, Summarizing}
\usage{
make_tidy_datetime_filename()
}
\arguments{
\item{datetime}{class: POSIXct}

\item{timezone}{class: string}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
make_tidy_datetime_filename(datetime=NA, timezone="UTC")
}
