% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_by.R
\name{mult_by}
\alias{mult_by}
\title{RM2C2: Scoring, Summarizing}
\usage{
mult_by(x, mult = 1)
}
\arguments{
\item{x}{class: numeric; original number}

\item{add}{class: numeric; number to add}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
mult_by(x, mult=5)
}
\keyword{cognition}
\keyword{m2c2,}
