% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_log_summary_export.R
\name{produce_log_summary_export}
\alias{produce_log_summary_export}
\title{RM2C2dev}
\usage{
produce_log_summary_export(
  filepath,
  part_ids = NA,
  search_events = NA,
  debug = F
)
}
\description{
RM2C2dev
}
