% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ambcog.R, R/stitch_datasets.R
\name{read_ambcog}
\alias{read_ambcog}
\alias{stitch_datasets}
\title{RM2C2: Scoring, Summarizing}
\usage{
read_ambcog(filepath, na = na)

stitch_datasets(df_list, common_id)
}
\arguments{
\item{filepath}{class: string}

\item{df_list}{class: list of dataframes}

\item{common_id}{class: vector of strings}
}
\description{
RM2C2: Scoring, Summarizing

RM2C2: Scoring, Summarizing
}
\examples{
read_ambcog("C:/filepath.csv")
stitch_datasets(df_list, common_id, outcomes = NULL)
}
\keyword{cognition}
\keyword{m2c2,}
