% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_m2c2_local.R
\name{read_m2c2_local}
\alias{read_m2c2_local}
\title{RM2C2dev}
\usage{
read_m2c2_local(filepath, na = na)
}
\arguments{
\item{filepath}{class: string}
}
\description{
RM2C2dev
}
\examples{
read_m2c2_local ("C:/filepath.csv")
}
\keyword{cognition}
\keyword{m2c2,}
