% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parsed_dir.R
\name{read_parsed_dir}
\alias{read_parsed_dir}
\title{RM2C2: Scoring, Summarizing}
\usage{
read_parsed_dir(
  dir,
  filetype = "txt",
  spattern = "gamedata_",
  recursive = T,
  na = "."
)
}
\arguments{
\item{filepath}{class: string}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
read_parsed_dir(dir, filetype='txt', spattern="gamedata_", recursive=T, na=".")
}
\keyword{cognition}
\keyword{m2c2,}
