% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_interference_data.R
\name{restructure_interference_data}
\alias{restructure_interference_data}
\title{RM2C2: Scoring, Summarizing}
\usage{
restructure_interference_data(df, nontap_cols = c(1:13), tap_col = c(14))
}
\description{
RM2C2: Scoring, Summarizing
}
