% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_tapping_data.R
\name{restructure_tapping_data}
\alias{restructure_tapping_data}
\title{RM2C2: Scoring, Summarizing}
\usage{
restructure_tapping_data(df, nontap_cols = c(1:17), tap_col = c(18))
}
\description{
RM2C2: Scoring, Summarizing
}
