% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dot_memory_alt.R
\name{score_dot_memory_alt}
\alias{score_dot_memory_alt}
\title{RM2C2: Scoring, Summarizing}
\usage{
score_dot_memory_alt(df, square_size = 5, n_dots = 3)
}
\description{
RM2C2: Scoring, Summarizing
}
