% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dot_memory_interference.R
\name{score_dot_memory_interference}
\alias{score_dot_memory_interference}
\title{RM2C2dev: Scoring, Summarizing}
\usage{
score_dot_memory_interference(
  df,
  nontap_cols = c(1:33, 35:37),
  tap_col = c(34)
)
}
\description{
RM2C2dev: Scoring, Summarizing
}
