% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_shopping_list.R
\name{score_shopping list}
\alias{score_shopping list}
\alias{score_shopping_list}
\title{RM2C2: Scoring, Summarizing}
\usage{
score_shopping_list(df, app_version = "1_1_release_18_7_27")
}
\description{
RM2C2: Scoring, Summarizing
}
