% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_adj.R
\name{SDT_adj}
\alias{SDT_adj}
\title{RM2C2: Scoring, Summarizing}
\usage{
SDT_adj(df, adj = 0.01)
}
\arguments{
\item{df}{class: dataframe}

\item{adj}{class: numeric; SDT adjustment}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
SDT_adj(df, adj=0.01)
}
\keyword{cognition}
\keyword{m2c2,}
