% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions_per_participant.R
\name{sessions_per_participant}
\alias{sessions_per_participant}
\title{RM2C2: Scoring, Summarizing}
\usage{
sessions_per_participant(df, id_var, session_var)
}
\arguments{
\item{df}{class: data frame}

\item{id_var}{class: string}

\item{session_var}{class: string}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
sessions_per_participant(df,"user_id","session_id")
}
\keyword{cognition}
\keyword{m2c2,}
