% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span_order.R
\name{span_order}
\alias{span_order}
\title{RM2C2: Scoring, Summarizing}
\usage{
span_order(df)
}
\arguments{
\item{df}{class: dataframe}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
span_order(df)
}
\keyword{cognition}
\keyword{m2c2,}
