% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyday_per_participant.R
\name{studyday_per_participant}
\alias{studyday_per_participant}
\title{RM2C2: Scoring, Summarizing}
\usage{
studyday_per_participant(df, id_var, studyday_var)
}
\arguments{
\item{df}{class: data frame}

\item{id_var}{class: string}

\item{session_var}{class: string}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
studyday_per_participant(df,"user_id","studyday")
}
\keyword{cognition}
\keyword{m2c2,}
