% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_change_detection.R
\name{summary_change_detection}
\alias{summary_change_detection}
\title{RM2C2: Scoring, Summarizing}
\usage{
summary_change_detection(df, group_var)
}
\description{
RM2C2: Scoring, Summarizing
}
