% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_color_shapes.R
\name{summary_color_shapes}
\alias{summary_color_shapes}
\title{RM2C2dev}
\usage{
summary_color_shapes(
  data,
  group_var,
  var_prefix = "color_shapes",
  experimental = F
)
}
\description{
RM2C2dev
}
