% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mot.R
\name{summary_mot}
\alias{summary_mot}
\title{RM2C2dev}
\usage{
summary_mot(
  data,
  group_var,
  var_prefix = "multi_obj_tracking",
  experimental = F,
  lure_label = "LURE",
  normal_label = "NORMAL"
)
}
\description{
RM2C2dev
}
