% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_seq_tapping.R
\name{summary_seq_tapping}
\alias{summary_seq_tapping}
\title{RM2C2: Scoring, Summarizing}
\usage{
summary_seq_tapping(df, group_var, nontap_cols = c(1:24))
}
\description{
RM2C2: Scoring, Summarizing
}
