% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_shopping_list.R
\name{summary_shopping_list}
\alias{summary_shopping_list}
\title{RM2C2: Scoring, Summarizing}
\usage{
summary_shopping_list(df, group_var, app_version = "1_1_release_18_7_27")
}
\description{
RM2C2: Scoring, Summarizing
}
