% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_symbol_search.R
\name{summary_symbol_search}
\alias{summary_symbol_search}
\title{RM2C2dev}
\usage{
summary_symbol_search(
  data,
  group_var,
  var_prefix = "symbol_search",
  basic_outcomes = T,
  lure_label = "LURE",
  normal_label = "NORMAL"
)
}
\description{
RM2C2dev
}
