% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_participants.R
\name{unique_participants}
\alias{unique_participants}
\title{RM2C2: Scoring, Summarizing}
\usage{
unique_participants(vec)
}
\arguments{
\item{vec}{class: vector of data}
}
\description{
RM2C2: Scoring, Summarizing
}
\examples{
unique_participants(df$user_id)
}
\keyword{cognition}
\keyword{m2c2,}
