% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_shoppinglist_items_across_phases.R
\name{verify_shoppinglist_items_across_phases}
\alias{verify_shoppinglist_items_across_phases}
\title{RM2C2dev}
\usage{
verify_shoppinglist_items_across_phases(filepath, debug = F)
}
\arguments{
\item{filepath}{class: string}

\item{debug}{class: boolean}
}
\description{
RM2C2dev
}
\examples{
verify_shoppinglist_items_across_phases(filepath, debug=T)
}
