% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_metadate_cols.R
\name{append_metadate_cols}
\alias{append_metadate_cols}
\title{RM2C2dev: Scoring, Summarizing}
\usage{
append_metadate_cols(
  data,
  time_var = "timestamp",
  time_format = "\%m/\%d/\%Y",
  time_origin = "1970-01-01",
  timezone = Sys.timezone(),
  append_rounded = T
)
}
\arguments{
\item{data}{class: data.frame}

\item{time_var}{class: string}

\item{time_format}{class: string}

\item{time_origin}{class: string}

\item{timezone}{class: string}

\item{append_rounded}{class: boolean}
}
\description{
RM2C2dev: Scoring, Summarizing
}
\examples{
append_metadate_cols(data, time_var = "timestamp")
}
