% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_metadate_cols.R
\name{append_metadate_cols}
\alias{append_metadate_cols}
\title{RM2C2dev}
\usage{
append_metadate_cols(
  data,
  time_var = "timestamp",
  time_format = "\%m/\%d/\%Y",
  time_origin = "1970-01-01",
  timezone = Sys.timezone(),
  append_rounded = T
)
}
\arguments{
\item{data}{class: data.frame}

\item{time_var}{class: string}

\item{time_format}{class: string}

\item{time_origin}{class: string}

\item{timezone}{class: string}

\item{append_rounded}{class: boolean}
}
\description{
RM2C2dev
}
\examples{
append_metadate_cols(data, time_var = 'timestamp')
parameter validation
verify data passed in is data.frame or tibble
append date features
pass in time_stamp as string and convert to date time object (as.POSIXct)
append date and time features
add weekend label
append rounded date timestamps if requested
append rounded date timestamps if requested
append rounded timestamps if requested
if data is valid, stop code execution and present an error message
}
