% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_process_cols.R
\name{append_process_cols}
\alias{append_process_cols}
\title{RM2C2dev}
\usage{
append_process_cols(data, algo = "md5")
}
\arguments{
\item{data}{class: data.frame}

\item{algo}{class: string}
}
\description{
RM2C2dev
}
\examples{
append_process_cols(data, algo='md5')
parameter validation
verify data passed in is data.frame or tibble
generate MD5 checksum of `data`
append MD5 checksum and file processing timestamp to `data`
add package version
if data is valid, stop code execution and present an error message
}
