% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_process_cols.R
\name{append_process_cols}
\alias{append_process_cols}
\title{RM2C2_dev: Scoring, Summarizing}
\usage{
append_process_cols(data, algo = "md5")
}
\arguments{
\item{data}{class: data.frame}

\item{algo}{class: string}
}
\description{
RM2C2_dev: Scoring, Summarizing
}
\examples{
append_process_cols(data, algo="md5")
}
