% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_frame_tibble.R
\name{is_data_frame_tibble}
\alias{is_data_frame_tibble}
\title{RM2C2_dev: Scoring, Summarizing}
\usage{
is_data_frame_tibble(data)
}
\arguments{
\item{data}{class: data.frame or tibble}
}
\description{
RM2C2_dev: Scoring, Summarizing
}
\examples{
is_data_frame_tibble(data)
}
