% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_dot_memory_interference.R
\name{score_dot_memory_interference}
\alias{score_dot_memory_interference}
\title{RM2C2dev: Scoring, Summarizing}
\usage{
score_dot_memory_interference(df)
}
\description{
RM2C2dev: Scoring, Summarizing
}
