% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_dot_memory_interference.R
\name{summary_dot_memory_interference}
\alias{summary_dot_memory_interference}
\title{RM2C2dev}
\usage{
summary_dot_memory_interference(
  data,
  group_var,
  var_prefix = "dot_memory",
  experimental = F
)
}
\description{
RM2C2dev
}
