% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_shopping_list.R
\name{summary_shopping_list}
\alias{summary_shopping_list}
\title{RM2C2dev}
\usage{
summary_shopping_list(
  data,
  group_var,
  var_prefix = "shopping_list",
  experimental = F,
  lure_label = "LURE",
  normal_label = "NORMAL"
)
}
\description{
RM2C2dev
}
